[Fixed] EmbeddedPaymentElement now errors when `configuration.rowSele… (#5135)
…ctionBehavior` is `.immediateAction` and
`configuration.embeddedViewDisplaysMandateText` is `true` to ensure
integrations display legal mandate text to the customer.

## Motivation
rowSelectionBehavior = immediateAction + embeddedViewDisplaysMandateText
= false is a broken integration.

1. When formSheetAction = continue...
...if you tap a `a PM row with no form and a mandate` you'll never see
the mandate that appears in embedded because you're immediately moved to
a different screen.

2. When formSheetAction = confirm...
...if you tap `a PM row with no form and a mandate` you'll never see the
mandate that appears in embedded because you will ~immediately (after a
network request or two) see a webview to authenticate.

## Testing
See unit test

## Changelog
See changelog