Fix Link inline signup to require phone number (#5164)
## Summary

Fixes an issue with Link inline signup where the primary CTA was enabled
when the Link signup checkbox was enabled, but no phone number was
provided. The phone number should be required in this scenario.

## Motivation

https://stripe.slack.com/archives/C08U9R2PY6N/p1752611466458249

## Testing

Before (Primary CTA is enabled when the phone number is empty): 

<img width=40%
src="https://github.com/user-attachments/assets/5725657f-c309-422c-8358-93376fee2efb"
/>

After (Primary CTA is disabled until the phone number is not empty):


https://github.com/user-attachments/assets/781f70e3-806a-47a6-89e2-f26abc503aaa

## Changelog

N/a