Fix CMTime related potential crash in SelfieUploader+API.swift (#5089)
## Summary

Validates the `exposureDuration` property before using it, to make sure
it is not infinite or NaN.

## Motivation

This should prevent a potential crash that is similar to one reported in
[https://github.com/stripe/stripe-ios/issues/3193](https://github.com/stripe/stripe-ios/issues/3193)

While investigating the above crash, we identified this line that is
similar to it and could cause crashes that are similar.

## Testing

Run through the identity flow and verify no crash occurs and it
completes successfully - specifically requiring selfie.

---------

Co-authored-by: Ken Ackerson <ken@lickability.net>