Add billingAddress to AddressElement (#5168)
## Summary
To gain parity with web we want to allow merchants to pass in a
billingAddress into AddressElement alongside a shipping address. When a
billing address is supplied by the merchant we will pre-fill it if no
shipping address is also provided. We will also show a checkbox "Use
billing address for shipping" that users can toggle to fill in the
address section with their billing address.

## Motivation
-
https://docs.google.com/document/d/1nLMw-zOGqhqyvjkAPdYQDRVPvCK17db6_s4ze7wHfaw/edit?tab=t.0

## Testing
- Updated test
- Manual

## Changelog
N/A