Extract Link signup controller from PayWithLinkViewController (#5139)
## Summary

In order to support a standalone Link authentication flow, we need to
extract the SignUpViewController / ViewModel out of
`PayWithLinkViewController`. The
`PayWithLinkViewController.SignUpViewController` still exists, it just
uses the extracted `LinkSignUpViewController` now.

In the next PR, I'll be adding support for a standalone Link
authentication flow, which makes use of this `LinkSignUpViewController`.

## Motivation

Standalone Link Auth

## Testing

Sign up flow still works:


https://github.com/user-attachments/assets/d3325e25-667c-4af8-9a1c-df0348fccc19



## Changelog

N/a