Tweak when Embedded didUpdateHeight is called (#5184)
## Summary
Before, `didUpdateHeight` was called when the _frame_ changed after
`layoutSubviews`.
Now, it's called when the "natural" height changed after
`layoutSubviews`.

This accommodates usage in e.g. a UITableViewCell where embedded must
inform its owner that the height changed _before_ it will be laid out
with a new frame.

Does this change the `didUpdateHeight` delegate contract?
> /// Called inside an animation block when the EmbeddedPaymentElement
view is updating its height. Your implementation should call
`setNeedsLayout()` and `layoutIfNeeded` on the scroll view that contains
the EmbeddedPaymentElement view. This enables a smooth animation of the
height change.

No, based on this description and intended usage, I don't think so.

## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->

## Testing
Manually tested the height change animation works as expected in:
- SwiftUI example 
- UIKit test playground
- A UITableView with a cell w/ embedded in it, that calls
`tableView.beginUpdates() / endUpdates()` in `didUpdateHeight`


https://github.com/user-attachments/assets/91e5d414-cb68-4fc9-a262-8ff76505716b

## Changelog
See changelog