Fix Identity Document Upload Popover iPad Crash (#5035)
## Summary

Fixes [Issue 2671](https://github.com/stripe/stripe-ios/issues/2671)

Fixes a crash on iPad that would occur in the native Identity flow when
attempting to select a document to upload. We were displaying an action
sheet in this case, but not specifying a `sourceView` or `sourceRect`
for the `popoverPresentationController`, causing the crash. Now, we pass
along the button when tapped and specify both to fix the crash.

## Testing

- Run the IdentityVerificationExample scheme on iPad simulator
- Go through the native flow until it fails to get camera
- Selected Upload a Photo to go to the upload screen and choose Select
- Verify it does not crash

---------

Co-authored-by: Ken Ackerson <ken@lickability.net>
Co-authored-by: Brian Capps <brian@lickability.net>