Fix error code bridging in Link (#5239)
## Summary
This code in Link wasn't working, as it was using the modern API
bindings. We should have a helper to bridge both modern and legacy error
codes.

## Motivation
Fixing a "an unknown error occurred" message when using Link.

## Testing
Added a test for the bridge, tested Link's error handling manually in PS
Example.

## Changelog
None