Allow skipping Link wallet for valid cards (#5245)
## Summary
<!-- Simple summary of what was changed. -->

This pull request adds the functionality to skip the Link wallet screen
and immediately return a valid payment method and shipping address to
the merchant if they're opening Link for payment method selection.

The functionality is currently gated behind
`link_mobile_skip_wallet_in_flow_controller`.

## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->

## Testing
<!-- How was the code tested? Be as specific as possible. -->
<!-- Ignored Tests: Did you newly ignore a test in this PR? If so,
please open an R4 incident so that the test can be re-enabled as soon as
possible-->

## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->

---------

Co-authored-by: Mat Schmid <mats@stripe.com>