Crypto Onramp SDK: Identity Verification Flow (#5272)
## Summary
Implements the rest of `promptForIdentityVerification()`, including use
of the Identity SDK to verify the user’s identity and return a result.

## Motivation
To implement the public API `promptForIdentityVerification` documented
[here](https://docs.google.com/document/d/1fiRBeegAqMbaR7OUDUy0EA4idvHEDh_FcBE2DTz2vyo/edit?tab=t.0).

## Testing
Added functionality to the example app to exercise the API. This video
shows success, failure, and cancelation results. Note that this also
works after registration, but the video only shows reaching this step
after verification.


https://github.com/user-attachments/assets/a212d7c3-43b4-424d-a35b-accff9d873ee

## Changelog
N/A