Update account onboarding component to accept new `requirements` param (#5279)
## Summary
Updates the account onboarding component to accept a new `requirements`
param. Allows the onboarding component to include only the requirements
in the `only` field or to `exclude` requirements in the `exclude` field.


https://github.com/user-attachments/assets/04742195-6f42-4119-91c9-2614b87aa60c

You can see in the video that when the param isn't set, onboarding will
collect all info. When it is set to `external_account`, it'll only ask
for that users's bank account.

## Motivation
These changes were previously approved in this [API
Review](https://jira.corp.stripe.com/browse/MOBILE_APIREVIEW-153).

## Testing
Manually tested by updating the example app. Updated one CI test as
well.

## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->