Adds `defaultFilters` option to the payments list component (#5287)
## Summary

BLUF, it allows you to select which filters apply on first component
load. The user can update them in the app afterwards. See the web
[docs](https://docs.stripe.com/connect/supported-embedded-components/payments#setting-default-filters)
for an explanation.

- Adds the `defaultFilters` prop to the payments component in the
Connect SDK.
- Updates the example app to let users input values.


https://github.com/user-attachments/assets/553f5028-9573-44a9-940f-1fa767de9765

This schema was approved in the [API
Review](https://docs.google.com/document/d/1ulycg5yokrMPDDpgWcsn2PtcRJHlp3KtmJeyF_PKVNI/edit?tab=t.0).

## Motivation
https://jira.corp.stripe.com/browse/CAX-4479

## Testing
<!-- How was the code tested? Be as specific as possible. -->
<!-- Ignored Tests: Did you newly ignore a test in this PR? If so,
please open an R4 incident so that the test can be re-enabled as soon as
possible-->

## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->