Crypto Onramp SDK: Adds NSCameraUsageDescription to Example App (#5297)
## Summary
Currently, if you follow the flow in the Identity SDK usage to preview
the end user experience, the example app will crash when attempting to
use camera APIs. This fixes that using the same camera description used
by the Identity Verification example.

## Motivation
To prevent a crash when using the example app.

## Testing
1. Authenticated with an existing user by using an OTP code in
CryptoOnramp Example.
2. Tapped the button to verify my identity.
3. Tapped "Proceed" on the identity test UI.
4. Tapped "Agree and Continue".
5. Tapped "I’m ready".
6. Verified that the app didn’t crash.

## Changelog
N/A