Support validation errors in Link UI (#5295)
## Summary
<!-- Simple summary of what was changed. -->

This pull request adopts the changes in
https://github.com/stripe/stripe-ios/pull/5290 for the native Link UI,
allowing us to show form validation errors if the user presses the
disabled primary button.

We use this in the add-card and update-card screens for now, as these
have the potential to grow quite big due to extensive billing details
collection.

## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->

## Testing
<!-- How was the code tested? Be as specific as possible. -->
<!-- Ignored Tests: Did you newly ignore a test in this PR? If so,
please open an R4 incident so that the test can be re-enabled as soon as
possible-->

## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->