Add `SellerDetails.businessName` field (#5317)
## Summary
<!-- Simple summary of what was changed. -->

This pull request adds a new `businessName` field to the
`SellerDetails`. This business name will be shown when paying with
wallets, while the `merchantDisplayName` will be shown in the reuse
mandate.

## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->

A request to be able to specify a different seller name from the
merchant ([Android
PR](https://github.com/stripe/stripe-android/pull/11339)).

## Testing
<!-- How was the code tested? Be as specific as possible. -->
<!-- Ignored Tests: Did you newly ignore a test in this PR? If so,
please open an R4 incident so that the test can be re-enabled as soon as
possible-->

## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->