[WalletButtonsView] Fix selecting the default PM (#5324)
## Summary
Don't allow Apple Pay as a valid default payment option when
WalletButtonsView is attached. Otherwise, we end up unsetting it later
(as Apple Pay is unavailable). Instead we'll either use the locally
saved PM default or fall through to the first saved PM.

## Testing
Tested in PS Example with WalletButtonsView and a returning customer.

## Changelog
Not a public API yet