Temporarily remove buildMenu override (#5332)
## Summary
The iOS 17 availability check appears to not work on Xcode 26 beta 5,
causing the example apps (and presumably other host apps) to crash
instantly on launch on devices/sim running iOS < 17. This change should
hopefully be able to be reinstated with future Xcode versions.

## Motivation
Avoid user app crashes when building with Xcode 26

## Testing
Doesn't crash!

## Changelog
N/A