Allow filtering supported payment method types in LinkController (#5309)
## Summary

This adds filtering the supported payment method types in
`LinkController`. It launches right into the financial connections flow
in the scenario where we've filtered to bank accounts, and no bank
accounts are available. I also added toggles in the example view to test
this out.

## Motivation


https://docs.google.com/document/d/1NL4a6Cbm5GdlWudaeL8dq8y0xZMvyIxv2CuaPbLYaXI/edit?usp=sharing

## Testing


https://github.com/user-attachments/assets/260c0cec-55b0-4656-b804-28a7e1eaf4e1

No existing bank accounts:


https://github.com/user-attachments/assets/0a6905e3-5f1d-4c0f-8a2e-472270a15a31

## Changelog

N/a