Crypto Onramp SDK: Appearance Adjustments for Keyboard Accessory Buttons (#5321)
## Summary

@mats-stripe helped to find another element that was still tinted with
Link brand colors: `DoneButtonToolbar`. However, this, and the elements
that use it, exist outside of `StripePaymentSheet` so we need to rely on
altering the themes (`ElementsAppearance`) that get passed across the
framework boundaries.

This is limited to affecting only the primary color of the theme, and
only in the case that the UI presentation originated from
`LinkController`, i.e. a `LinkAppearance` instance exists.

## Motivation
To allow clients to theme Link components with their own branding
colors.

## Testing

Tapped every field accessible in the UI presented from the CryptoOnramp
Example project’s payment selector, and verified that in each case, the
"Done" button above the keyboard was the correct lavender color.


https://github.com/user-attachments/assets/fbf55a7f-1a9b-4354-9a4a-72b3186de04c

Additionally, I forced some UI components to appear that are either
permanently hidden, or shown under certain circumstances, to ensure they
were the right color. This screenshot shows the checkbox element, and
the card brand selection element which also includes a Done button.

<img
src="https://github.com/user-attachments/assets/0aa18f99-5c05-4099-92eb-71237666739c"
width="300" />

## Changelog

N/A