Fix when client_attribution_metadata is sent (#5366)
## Summary
<!-- Simple summary of what was changed. -->
API bindings were incorrectly sending client attribution metadata
because the clientAttributionMetadata on the PaymentMethodParams had a
default value. Changed the default value to be nil
## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->
Project Weave

https://docs.google.com/document/d/1NNTf3TKRJ6Ftgk6PYtDjPMSAYU-RMUezUjGndJr81Eg/edit?pli=1&tab=t.0#heading=h.8vkqyui1b9gs
## Testing
<!-- How was the code tested? Be as specific as possible. -->
<!-- Ignored Tests: Did you newly ignore a test in this PR? If so,
please open an R4 incident so that the test can be re-enabled as soon as
possible-->
Updated network stubs
## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->
N/A