Crypto Onramp SDK: Example App Attach Wallet Address Sheet (#5377)
## Summary
Adds an interface to attach a wallet address that calls the
coordinator’s `collectWalletAddress` method. Note that the wallet
information is not stored or persisted locally, or fetched the next time
this UI appears. Defaults to a test eth wallet address that we’re
currently using.

## Motivation
Working towards an end-to-end checkout example.

## Testing


https://github.com/user-attachments/assets/3758bf15-8152-4e93-8169-ff57a9610735

## Changelog

N/A