Fix Link inline signup issues (#5384)
## Summary
<!-- Simple summary of what was changed. -->

This pull request fixes two issues with the Link signup opt-in feature:
1. It preserves the checkbox state when re-opening the card form. This
was broken because `updateForWalletButtons` recreated the card form with
the initial checkbox state, not the most recent one.
2. It ensures that we always send a `country_inferring_method` in the
signup call.

<details><summary>Recording</summary>
<p>


https://github.com/user-attachments/assets/9545ad7b-ed3a-4a9e-91af-a479f01e5f86

</p>
</details> 

## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->

## Testing
<!-- How was the code tested? Be as specific as possible. -->
<!-- Ignored Tests: Did you newly ignore a test in this PR? If so,
please open an R4 incident so that the test can be re-enabled as soon as
possible-->

## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->