Show Link reauthentication inline if applicable (#5404)
## Summary
<!-- Simple summary of what was changed. -->

This pull request updates how we re-authenticate Link consumers if their
session has expired before re-opening the wallet.

Previously, we used the `retryingOnAuthError` method to have consumers
verify their account when we encountered an expired session while
loading the wallet. This caused a suboptimal UI (a loading dialog on top
of the loading sheet) and could result in an endless loading state due
to a race condition.

With this pull request, we no longer handle re-authentication for each
API call separately. Instead, we catch any authentication-related errors
in `loadAndPresentWallet` and navigate to the verification screen
_after_ creating a new consumer session, which causes both payment
details and shipping addresses to be fetched with the new consumer
session client secret.

## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->

## Testing
<!-- How was the code tested? Be as specific as possible. -->
<!-- Ignored Tests: Did you newly ignore a test in this PR? If so,
please open an R4 incident so that the test can be re-enabled as soon as
possible-->

## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->