CryptoOnramp SDK: `performCheckout`  Adjustments (#5427)
## Summary
While working on the ReactNative SDK, we discovered that the Android
implementation of `performCheckout` has a "canceled" result indicating
that the user canceled the checkout (i.e. 3DS was dismissed).

This PR brings iOS in line with this, while also `throw`ing errors in
the case failure instead of returning a `.failed` enum value.

## Motivation
To allow the ReactNative SDK implementation of CryptoOnramp to emit the
same checkout results on both platforms.

## Testing
Verified that I was still able to make it through the checkout flow.
<img width="300" alt="Simulator Screenshot - iPhone 16 Pro - 2025-08-27
at 16 42 20"
src="https://github.com/user-attachments/assets/fc763081-fc0f-4b42-bd9b-fb9b6f26ec6d"
/>

## Changelog
N/A

---------

Co-authored-by: Mat Schmid <mats@stripe.com>