Pass `allow_redisplay` along to FC SDK (#5437)
## Summary
<!-- Simple summary of what was changed. -->

This pull request makes us pass along the merchant-provided
`allow_redisplay` value to the Financial Connections SDK. The Stripe.js
implementation will be
[updated](https://git.corp.stripe.com/stripe-internal/pay-server/pull/1180553)
to then pass this value in the payment method creation call.

## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->

## Testing
<!-- How was the code tested? Be as specific as possible. -->
<!-- Ignored Tests: Did you newly ignore a test in this PR? If so,
please open an R4 incident so that the test can be re-enabled as soon as
possible-->

## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->