Attempt to fix potential race condition in image scanning session's concurrency manager  (#5055)
## Summary

Tracks the balance of a semaphore in the scanning session concurrency
code. If that is unbalanced by the time it goes out of memory, it will
resolve it. I do not see a way this crash happens, but this should
resolve it in any circumstance.

## Motivation

Fixes [#2670](https://github.com/stripe/stripe-ios/issues/2670)

## Testing

I went through the identity flow a handful of times, and cancelling mid
flow as well during scanning.

---------

Co-authored-by: Ken Ackerson <ken@lickability.net>